Changed math classes to templates with typedefs. Added both UDP and TCP sockets and other Socket utils. Added Memory stream to assist with serialization and began work on a reflection system.

+++ CubeWave/src/Wave.cpp
@@ -16,11 +16,11 @@ namespace CubeWave
 
 	void Wave::Update(double elapsedSeconds)
 	{
-		Matrix4d& transform = m_GameObject->mesh().GetTransform(0);
+		Matrix4f& transform = m_GameObject->mesh().GetTransform(0);
 		float distance = Vector2f::Distance(m_GameObject->transform().Position().xz(), m_Center->transform().Position().xz());
 		float offset = Map(distance, 0, m_MaxDist, 0, m_Offset);
 		float yScale = Map(sin(m_Angle + offset), -1.0f, 1.0f, m_Min, m_Max);
-		transform = Matrix4d::Scale(1, yScale, 1);
+		transform = Matrix4f::Scale(1, yScale, 1);
 		m_Angle += m_Speed * Time::DeltaTime();
 	}
